refactor: move type definitions for auth
