+++ packages/rest-api-client/src/types/auth.ts
@@ -1,0 +1,30 @@
+type ApiTokenAuth = {
+  type: "apiToken";
+  apiToken: string | string[];
+};
+
+type PasswordAuth = {
+  type: "password";
+  username: string;
+  password: string;
+};
+
+type SessionAuth = {
+  type: "session";
+};
+
+type OAuthTokenAuth = {
+  type: "oAuthToken";
+  oAuthToken: string;
+};
+
+export type DiscriminatedAuth =
+  | ApiTokenAuth
+  | PasswordAuth
+  | SessionAuth
+  | OAuthTokenAuth;
+
+export type BasicAuth = {
+  username: string;
+  password: string;
+};
